requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0